+++ configuration.conf
@@ -30,7 +30,7 @@ vrp_basic           | N | Y | Y
 withPoints          | Y | Y | Y
 lineGraph           | Y | Y | Y
 components          | Y | Y | Y
-bellman_ford        | Y | Y | Y
+bellman_ford        | Y | Y | N
 cpp_common          | Y | N | N
 dagShortestPath     | Y | Y | Y
 chinese             | Y | Y | Y
@@ -39,9 +39,7 @@ mincut              | Y | Y | Y
 version             | Y | Y | Y
 topologicalSort     | Y | Y | Y 
 transitiveClosure   | Y | Y | Y
-breadthFirstSearch  | Y | Y | N
-binaryBreadthFirstSearch    | Y | Y | Y  
-edwardMoore         | Y | Y | N
+breadthFirstSearch  | Y | Y | Y
 #----------------------
 # SQL only directories
 #----------------------
