+++ src/breadthFirstSearch/CMakeLists.txt
@@ -1,4 +1,6 @@
 ADD_LIBRARY(breadthFirstSearch OBJECT
     breadthFirstSearch.c
     breadthFirstSearch_driver.cpp
+    binaryBreadthFirstSearch.c
+    binaryBreadthFirstSearch_driver.cpp    
     )
